+++ packages/terra-core-docs/src/terra-dev-site/doc/section-header/example/ClosedSectionHeader.jsx
@@ -6,6 +6,6 @@ const sectionHeaderProps = {
   onClick: () => {},
 };
 
-const ClosedSectionHeader = () => <SectionHeaderExampleTemplate title="Closed Section Header" exampleProps={sectionHeaderProps} />;
+const ClosedSectionHeader = () => <SectionHeaderExampleTemplate text="Closed Section Header" exampleProps={sectionHeaderProps} />;
 
 export default ClosedSectionHeader;