+++ packages/terra-core-docs/src/terra-dev-site/test/action-header/CloseActionHeader.test.jsx
@@ -4,7 +4,7 @@ import ActionHeader from 'terra-action-header/lib/ActionHeader';
 
 const ActionHeaderExample = () => (
   <ActionHeader
-    title="Close Action Header"
+    text="Close Action Header"
     onClose={() => alert('You clicked close!')}
   />
 );