+++ packages/terra-core-docs/src/terra-dev-site/test/action-header/MinimizeCustomButtonActionHeader.test.jsx
@@ -4,7 +4,7 @@ import Button from 'terra-button';
 import ActionHeader from 'terra-action-header/lib/ActionHeader';
 
 const ActionHeaderExample = () => (
-  <ActionHeader title="Minimize Custom Button Action Header" onMinimize={() => alert('You clicked minimize!')}>
+  <ActionHeader text="Minimize Custom Button Action Header" onMinimize={() => alert('You clicked minimize!')}>
     <Button text="Custom Button" onClick={() => alert('You clicked me!')} />
   </ActionHeader>
 );