+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+## 1.23.0 - (March 1, 2023)
+
+* Changed
+  * Updated action-header examples to include A11y labels.
+  * Updated `terra-button-group` tests and examples.
+
+* Added
+  * Added accessibility guide for `terra-button-group`.
+
 ## 1.22.0 - (February 16, 2023)
 
 * Changed